basic session validation for pages
