+++ src/AwesomeApp.UI/src/pages/api/auth/[...nextauth].tsx
@@ -1,4 +1,4 @@
-import { nextAuthOptions } from "@/shared/libs";
+import { nextAuthOptions } from "@/backend/libs";
 import NextAuth from "next-auth";
 
 export default NextAuth(nextAuthOptions)
