+++ src/AwesomeApp.UI/src/pages/create-account/index.tsx
@@ -1,4 +1,4 @@
-import { ReactElement, useCallback, useEffect, useMemo, useState } from "react";
+import { ReactElement } from "react";
 import { PageBox } from "@/frontend/components";
 import { Button, FormControl, FormControlLabel, FormLabel, Grid, Radio, RadioGroup, TextField, Typography } from "@mui/material";
 import Link from "next/link";
