--- src/AwesomeApp.UI/src/shared/libs/SessionHelper.ts
@@ -1,16 +1,0 @@
-import { nextAuthOptions } from "@/backend/libs"
-import { NextApiRequest, NextApiResponse } from "next"
-import { getServerSession } from "next-auth"
-
-export async function getSession(req: NextApiRequest, res: NextApiResponse) {
-  const session = await getServerSession(req, res, nextAuthOptions)
-
-  if (!session || !session.user) {
-    throw new Error("Not authorized")
-  }
-
-  return {
-    ...session,
-    user: session.user!
-  }
-}
