+++ src/AwesomeApp.UI/src/shared/libs/index.ts
@@ -1,4 +1,3 @@
 export * from './ActionResultFactories'
 export * from './AxiosClientHelper'
 export * from './EnvironmentHelpers'
-export * from './SessionHelper'
